REM-156: Update transfer
