+++ remme/tp/account.py
@@ -41,6 +41,12 @@ class AccountHandler(BasicHandler):
     def __init__(self):
         super().__init__(FAMILY_NAME, FAMILY_VERSIONS)
 
+    @classmethod
+    def create_transfer(cls, context, address_from, address_to, amount):
+        transfer_payload = TransferPayload(address_to=address_to, value=amount)
+        return cls() \
+            ._transfer_from_address(context, address_from, transfer_payload)
+
     def get_state_processor(self):
         return {
             AccountMethod.TRANSFER: {
@@ -82,3 +88,52 @@ class AccountHandler(BasicHandler):
             raise InvalidTransaction("Receiver address has to be of "
                                      "an account type")
         return self._transfer_from_address(context, address, transfer_payload)
+
+    def _transfer_from_address(self, context, address, transfer_payload):
+        signer_key = address
+
+        if not transfer_payload.value:
+            raise InvalidTransaction("Could not transfer with zero amount")
+
+        if not transfer_payload.address_to.startswith(self._prefix) \
+                and transfer_payload.address_to not in [ZERO_ADDRESS]:
+            raise InvalidTransaction("Receiver address has to be of an account type")
+
+        if signer_key == transfer_payload.address_to:
+            raise InvalidTransaction("Account cannot send tokens to itself.")
+
+        signer_account, receiver_account = get_multiple_data(context, [
+            (signer_key, Account),
+            (transfer_payload.address_to, Account)
+        ])
+
+        # TODO transfer from genesis address using SETTINGS_KEY_GENESIS_OWNERS
+        # list of allowed addresses(0x0)
+        # genesis_members_str = _get_setting_value(context,
+        #                                          SETTINGS_KEY_GENESIS_OWNERS)
+        # if not genesis_members_str:
+        #     raise InvalidTransaction('REMchain is not configured '
+        #                              'to process genesis transfers.')
+        #
+        # genesis_members_list = genesis_members_str.split()
+
+        if not receiver_account:
+            receiver_account = Account()
+        if not signer_account:
+            signer_account = Account()
+
+        if signer_account.balance < transfer_payload.value:
+            raise InvalidTransaction(
+                "Not enough transferable balance. Sender's current balance: "
+                f"{signer_account.balance}")
+
+        receiver_account.balance += transfer_payload.value
+        signer_account.balance -= transfer_payload.value
+
+        LOGGER.info(f'Transferred {transfer_payload.value} tokens from '
+                    f'{signer_key} to {transfer_payload.address_to}')
+
+        return {
+            signer_key: signer_account,
+            transfer_payload.address_to: receiver_account
+        }
