+++ tests/tp/test_account.py
@@ -125,6 +125,16 @@ class AccountTestCase(HelperTestCase):
         self.expect_ok()
 
     @test
+    def test_transfer_fail_to_oneself(self):
+        ACCOUNT_AMOUNT1 = 500
+        TRANSFER_VALUE = 200
+        self.send_transaction(AccountMethod.TRANSFER,
+                              AccountClient.get_transfer_payload(self.account_address1, TRANSFER_VALUE),
+                              [self.account_address1, self.account_address2])
+        # self.expect_get({self.account_address1: AccountClient.get_account_model(ACCOUNT_AMOUNT1)})
+        self.expect_invalid_transaction()
+
+    @test
     def test_transfer_fail_to_zeroaddress(self):
         TRANSFER_VALUE = 200
         self.send_transaction(AccountMethod.TRANSFER,
